% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osconfig.R
\docType{class}
\name{bluemix-class}
\alias{bluemix}
\alias{bluemix-class}
\title{sparkcontext:  a SparkContext object.}
\description{
credentials:  a dictionary with the following required keys:
  
  auth_url
  project_id (or projectId)
  user_id (or userId)
  password
  region
and optional key:
  name  #[to be deprecated] The name of the configuration.
name:  string that identifies this configuration. You can
    use any string you like. This allows you to create
    multiple configurations to different Object Storage accounts.
    This is not required at the moment, since credentials['name']
    is still supported.
When using this from a IBM Spark service instance that
is configured to connect to particular Bluemix object store
instances, the values for these credentials can be obtained
by clicking on the 'insert to code' link just below a data
source.
}
\examples{
library(ibmos2spark)
creds = list(name="XXXXX", 
             auth_url="https://identity.open.softlayer.com",
            region="dallas", 
            project_id = "XXXXX", 
            user_id="XXXXX", 
            password="XXXXX",
            public = FALSE)
            
bmsc = bluemix(sparkcontext=sc, name=name, credentials = creds)

data <- read.df(sqlContext, bmsc$url(name, space,object), source = "com.databricks.spark.csv", header = "true")
head(data)

}

