% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osconfig.R
\docType{class}
\name{softlayer2d-class}
\alias{softlayer2d}
\alias{softlayer2d-class}
\title{sparkcontext is a SparkContext object.}
\description{
name is a string that identifies this configuration. You can
    use any string you like. This allows you to create
    multiple configurations to different Object Storage accounts.
auth_url, username and password are string credentials for your
Softlayer Object Store
}
\examples{
library(ibmos2spark)
slsc = softlayer2d(sparkcontext=sc, 
                 name="XXXXX", 
                 auth_url="https://identity.open.softlayer.com",
                 region="XXXXX", 
                 tenant = "XXXXX", 
                 username="XXXXX", 
                 password="XXXXX"
           )
           
data <- read.df(sqlContext, slsc$url(name, space,object), source = "com.databricks.spark.csv", header = "true")
head(data)

}

